pyproject.toml
[project]
name = "django-inertia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-tailwindcss>=0.4.0",
    "django-vite>=3.1.0",
    "inertia-django>=1.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.11.0",
]
